Module rymdbiljard::arrow

Reexports

use piston::input::RenderArgs;

use opengl_graphics::GlGraphics;

use graphics::Transformed;

use graphics::line;

use na::{Norm, Point2, Vector1, Vector2, Rotation2, Rotate};

use std::f64;

Structs

Arrow
Basic arrow entity which is used for drawing the direction and power in the shooting mode

Enums

ShootingMode
Mode to distinguish between the different staged of user interaction with the white poolball. In rotate mode the used decides the direction whereas in Power mode the user decides the poolball's initial speed

Constants

DEFAULT_LENGTH

MAX_LENGTH

ROTATION_SPEED