Struct rymdbiljard::blackhole::Blackhole

pub struct Blackhole {
    pub position: Point2<f64>,
    pub mass: f64,
    pub radius: f64,
    pub reach: f64,
}

Blackhole contains information about a single blackhole in the game, and methods for rendering and interacting with other entities

Fields
position: Point2<f64> mass: f64 radius: f64 reach: f64
Methods

impl Blackhole

fn new(position: Point2<f64>, mass: f64, radius: f64, reach: f64) -> Blackhole

Constructs a new Blackhole with the given properties

fn is_spagettified(&self, poolball: &Poolball) -> bool

Returns true if the given poolball is in contact with the blackhole

fn render(&self, args: &RenderArgs, gl: &mut GlGraphics)

Renders itself using the given graphics