Module rymdbiljard::game

Reexports

use piston::input::*;

use opengl_graphics::GlGraphics;

use opengl_graphics::glyph_cache::GlyphCache;

use na::Point2;

use std::f64;

use std::process::exit;

use poolball;

use goalzone;

use blackhole;

use physics;

use arrow;

Structs

CollisionPair
Struct used for holding information about a ball-ball collision or a ball-wall collision. A ball-wall collision is signified by Second being none.

Game
Contains information about the global game state as well as methods for handling the overarching game mechanics including the update loop and the collision handling algorithm.

Functions

white_ball_position
Searches for the white ball in the vector returning its index