Struct rymdbiljard::game::Game

pub struct Game {
    balls: Vec<Poolball>,
    blackholes: Vec<Blackhole>,
    goalzones: Vec<Goalzone>,
    score: i32,
    arrow: Arrow,
}

Contains information about the global game state as well as methods for handling the overarching game mechanics including the update loop and the collision handling algorithm.

Fields
balls: Vec<Poolball> blackholes: Vec<Blackhole> goalzones: Vec<Goalzone> score: i32 arrow: Arrow
Methods

impl Game

fn new(balls: Vec<Poolball>,
blackholes: Vec<Blackhole>,
goalzones: Vec<Goalzone>)
-> Self

Creates a new game given the specified parameters

fn render(&mut self,
gl: &mut GlGraphics,
args: &RenderArgs,
cache: &mut GlyphCache)

Renders the current game state including the poolballs, current score, blackholes, goalzones and the arrow indicator using GlGraphics

fn try_switch_mode(&mut self)

Attemps to switch to the next mode in the shooting stage doing so if and only if the white ball is stationary

fn update(&mut self, args: &UpdateArgs)

Updates the positon, speeds and so on for all poolballs aswell as handling the collisions

fn get_first_collision_pair(&self) -> CollisionPair

Returns a collision pair for the earlies collision by going throguh all poolballs searching for the ball-wall or ball-ball pair with the earlies collision time