Struct rymdbiljard::goalzone::Goalzone

pub struct Goalzone {
    position: Point2<f64>,
    radius: f64,
}

Implements a basic goalzone to be used in each corner of the playing field.

Fields
position: Point2<f64> radius: f64
Methods

impl Goalzone

fn new(position: Point2<f64>) -> Goalzone

Creates a new goalzone at the given position

fn reached_goal(&self, ball: &Poolball) -> bool

Returns true if the given poolball is inside the goalzone

fn render(&self, args: &RenderArgs, gl: &mut GlGraphics)

Renders itself using the given graphics