Module rymdbiljard::physics

Reexports

use na::{Vector2, Point2, Norm, FloatPoint, Dot};

use poolball;

use blackhole;

use math;

use std::f64;

Constants

FRICTION

Functions

ball_ball_collision
Calculates the new velocities for 2 colliding poolballs after time delta_time, (delta_time is in seconds)

ball_wall_collision
Recalculates the new velocities for the ball given collision with a wall

calculate_friction
Calculates the direction and size of the friction acceleration on the given ball

calculate_gravity
Calculates the acceleration acting upon the ball from the given black holes

gravity_acceleration
Calculates the gravity acceleration from an object with given mass at given distance

time_to_ball_ball_collision
Returns the time until the given two poolballs collide given their current velocities. Infinity is returned if the two poolballs never collide

time_to_wall_collision
Return the time to impact with wall given the current velocity