Struct rymdbiljard::poolball::Poolball

pub struct Poolball {
    pub position: Point2<f64>,
    pub radius: f64,
    pub mass: f64,
    pub velocity: Vector2<f64>,
    pub ball_type: BallType,
}

Poolball containing information about a single poolball in the game, and methods for manipulating this poolballs physical status

Fields
position: Point2<f64> radius: f64 mass: f64 velocity: Vector2<f64> ball_type: BallType
Methods

impl Poolball

fn new(position: Point2<f64>, ball_type: BallType) -> Poolball

Creates a new Poolball with the given initial position and velocity

fn update(&mut self, delta_time: f64)

Updates the poolball position using its current velocity

fn update_velocity(&mut self, acceleration: Vector2<f64>, delta_time: f64)

Updates the velocity of the poolball given the perceived acceleration

fn set_velocity(&mut self, new_velocity: Vector2<f64>)

Sets the velocity of the ball to the given velocity

fn get_value(&self) -> i32

Returns the score value of the poolball based on the ball type - White: -100 - Red: 10 - Blue: 30

fn is_stationary(&self) -> bool

Returns true if the poolball is stationary

fn render(&self, args: &RenderArgs, gl: &mut GlGraphics)

Renders itself using the given graphics